Use any host that mounts the datastore to create Volume
Also, This change makes zone to work per datacenter and cleans up dummy vms.
There can be multiple datastores found for a given name. The datastore name is
unique only within a datacenter. So this commit returns a list of datastores
for a given datastore name in FindDatastoreByName() method. The calles are
responsible to handle or find the right datastore to use among those returned.

Kubernetes-commit: 18922a3c44874bd052f06a2a9daae58da935f797