Fix closing of dirs in doSafeMakeDir
This fixes the issue where "childFD" from syscall.Openat is assigned to
a local variable inside the for loop, instead of the correct one in the
function scope. This results in that when trying to close the "childFD"
in the function scope, it will be equal to "-1", instead of the correct
value.

Kubernetes-commit: 07bbc7527a36141d3c0a9162c6f73cc5cba3d433