Fix attachment of just detached AWS volumes
AWS API has eventual consistency model and it can return stale data.
For example, DescribeInstances can return that a volume is still being
detached, while it has already been confirmed to be detached.

AttachDisk() should return error in this case, allowing A/D controller
to retry, and not to assume that the volume is attached.

Kubernetes-commit: 4c0c97d0db77aee06a654a4113b4e7422111895e