vsphere: check if volume exists before create
Sometimes, volume creation can succeed right as the request times out,
causing k8s to interpret it as a failure. When the request is retried,
we want it to succeed. When trying this in vSphere, the second create
request failed with "already exists" and it never recovered.

This adds a check to the in-tree vsphere storage plugin that checks if a
VMDK exists before trying to create it. The check is done BEFORE create.

Tested: manual only )-:

Kubernetes-commit: d84696ea54f771dd8a4d7517b75b39fe3b80f49c