Azure: Filter disks with ToBeDetached flag from attach/detach
After the detach calls are made, any further updates with the disks in
the process of getting detached will complain with errors such as:

"
Cannot attach data disk '<disk name>' to VM '<vm name>' because the
disk's ToBeDetached flag is set to true. Please either remove the
disk from APIModel or set the ToBeDetached flag to false and then
try again
"

This PR will ensure that the disks with the ToBeDetached flag set are
filtered out before the update to the cloud is made.

Kubernetes-commit: c6b8207bfe508995a5b641c21acae26a878512be