Fix AWS eventual consistency of AttachDisk
AWS eventual consistency can go back in time. It can return that a volume
is detached and then that it is attached.

When this happens during attachment of the same volume to the same node,
but with a different device name, retry DescribeVolumes a few times before
reporting an error. 10 retries should be enough to get a consistent result.

Kubernetes-commit: 4dd6fd448060b5a1ae29ed3bf766459f03963b2c