Fix build when `providerless` tags is provided
If someone tries to compile int providerless mode / with the build tag
'providerless', some bits are excluded and compilation will fail. This
is due to the fact that in this release we don't support providerless
yet.
Therefore we should remove the build tags from the files to not
accidentaly exclude them.

Kubernetes-commit: cd7cb006e88948f6c26809d2601f626c7da0a613