Don't report deletion of attached volume as warning
It is part of normal cluster operation. When a pod and PVC are deleted at
the same time (e.g. because of the whole namespace was deleted), PV
controller may try to delete a volume that is still detached.

"Warning" is too strong here, it is going to heal relatively quickly.

Kubernetes-commit: ee85b6a579f4d2c45705038c766833ecfc212a66