Do not skip externalLB update if some nodes are not found.
Log a warning instead and continue with the update. This is useful in cases where
the number of nodes is changing due to autoscaling or updgrades. It is possible
that the nodes picked by service controller don't all exist when gce layer lists
them. Update should still succeed with the nodes in the input that are valid.
This will still return an error if 0 nodes were found, when a non-zero input was passed in.

same

Kubernetes-commit: c1597beb15416efc789be9a2ff9aef6e84294ef2