Add node check to vSphere cloud provider
At KCM startup, vSphere cloud provider builds a cache from NodeAdded events
from an informer. But these events are asynchronous, the cloud provider may
need information from the cache earlier, for example when detaching a
volume from a node at KCM startup. If a node is missing in the cache,
the cloud provider treats such a detach as successful, which is wrong.
Such a volume will be attached to the node forever.

To prevent this issue:

1. Try nodeLister before declaring a node as not found.
A/D controller starts after its node informer has been synced.

2. Read API server before declaring a node as not found.
Just in case the informer has stale data.

Kubernetes-commit: fbcb7181f2c34e5524e19e5345d90c51b960cdfe