gce: fix test for non-existing instance in cloud
This commit fixes a GCE instance test for a scenario where instance does
not exist in the cloud. In the `gce_instances.go` we have a code that
handles such a scenario and strips off the error, i.e.

```
if providerID, err = cloudprovider.GetInstanceProviderID(ctx, g, types.NodeName(node.Name)); err != nil {
	if err == cloudprovider.InstanceNotFound {
		return false, nil
	}
	return false, err
}
```

but nevertheless the test was expecing a non-empty error string. This
issue got exposed when implementation of `GetInstanceProviderID` in the
`cloud-provider/cloud.go` changed to return `InstanceNotFound` as-is and
not as a custom string.

Signed-off-by: Mat Kowalski <mko@redhat.com>

Kubernetes-commit: da55d879b9e497f8941762b94b1d0d32f17fb6b8